+++ _evan/Musubi/CHANGELOG.md
@@ -2,3 +2,8 @@
 ## [1.0.0] - 8/28/2023
 ### Changed
 * First stable build
+## [1.1.0] - 8/30/2023
+### Fixed
+* Removed all manner of unwanted clicking on fast attacks
+### Added
+* Volume meter display